Changed interface files and moved all macros to same file
