+++ include/interface.h
@@ -1,0 +1,11 @@
+#ifndef INTERFACE
+#define INTERFACE
+
+/*************************************************************/
+/*                    FUNCTION PROTOTYPES                    */
+/*************************************************************/
+
+void show_cin_info();
+void show_prompt();
+
+#endif
