+++ include/macros.h
@@ -5,8 +5,17 @@
 /*                      DEFINE SECTION                       */
 /*************************************************************/
 
-#define MAX_STRING_SIZE 1024
+#define CIN_VERSION     "v1.2.3"
+
 #define FILENAME_C      "/tmp/cin/tmp_code.c"
 #define OUTPUT          "/tmp/cin/out.app"
+
+#define MAX_STRING_SIZE 1024
+#define INCLUDE_STR      "#include"
+#define INCLUDE_STR_SIZE 8
+#define DEFINE_STR       "#define"
+#define DEFINE_STR_SIZE  7
+
+#define MAIN_SIGNATURE  "int main(int argc, char**argv) {\n"
 
 #endif
